aggiunto allert di conferma quando eseguiamo una transazione importante come una eliminazione

+++ src/main/java/oop/focus/finance/view/windows/TransactionDetailsWindowImpl.java
@@ -1,5 +1,6 @@
 package oop.focus.finance.view.windows;
 
+import javafx.scene.control.ButtonType;
 import oop.focus.finance.controller.FXMLPaths;
 import oop.focus.finance.controller.TransactionsController;
 import oop.focus.finance.model.Transaction;
@@ -15,7 +16,10 @@ public class TransactionDetailsWindowImpl extends GenericDetailsWindow<Transacti
 
     @Override
     public final void save() {
-        super.getController().deleteTransaction(super.getX());
+        var result = super.confirm("Sicuro di voler elminare la transazione?");
+        if (result.isPresent() && result.get() == ButtonType.OK) {
+            super.getController().deleteTransaction(super.getX());
+        }
         this.close();
     }
 }
